Add CRAM runner and simple test
