+++ src/meson.build
@@ -3,13 +3,19 @@ hmcnc_cpp_sources = files([
   'viterbi.cpp',
 ])
 
- executable(
+hmcnc_cpp_sources = files([
+  'hmmcnc.cpp',
+  'hmcnc_io.cpp',
+])
+
+hmcnc_exe_sources = files([
+  'main.cpp',
+])
+hmcnc_exe_sources += hmcnc_cpp_sources
+
+hmcnc_exe = executable(
   'hmcnc',
-  files([
-    'hmmcnc.cpp',
-    'hmcnc_io.cpp',
-    'main.cpp',
-  ]),
+  hmcnc_exe_sources,
   install : true,
   dependencies : hmcnc_lib_deps,
   cpp_args : hmcnc_cpp_args,
