+++ tests/data/chr22.fa.fai
@@ -1,0 +1,1 @@
+chr22	50818468	125	70	71
