+++ tests/meson.build
@@ -1,25 +1,2 @@
-hmcnc_tests_cpp = files([
-  'test_hmcnc.cpp',
-  'test_hmcnc_io.cpp',
-  '../src/hmmcnc.cpp',
-  '../src/hmcnc_io.cpp',
-])
-
-hmcnc_gtest_dep = dependency(
-  'gtest',
-  main: true,
-  fallback: ['gtest', 'gtest_main_dep']
-)
-
-hmcnc_test = executable(
-  'hmcnc_test',
-  hmcnc_tests_cpp,
-  dependencies : [hmcnc_lib_deps, hmcnc_gtest_dep],
-  cpp_args : hmcnc_cpp_args,
-  install : false
-)
-
-test(
-  'hmcnc unit tests',
-  hmcnc_test,
-)
+subdir('unit')
+subdir('cram')
