+++ tests/unit/test_hmcnc.cpp
@@ -1,0 +1,7 @@
+#include <gtest/gtest.h>
+
+#include "hmcnc.h"
+
+TEST(hmcnc, test_setup_check) {
+  EXPECT_EQ(42, hmcnc_test());
+}
